ai_work/requirements.txt
pika
minio
ultralytics
fastapi
uvicorn
requests
jpype1